requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.2.4
django-admin-interface==0.16.3
django-admin-rangefilter==0.8.1
django-advanced-filters==1.3.0
django-braces==1.14.0
django-colorfield==0.4.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-simplefilters==0.1.3
docopt==0.6.2
filelock==3.0.12
idna==2.10
multidict==5.1.0
pipreqs==0.4.10
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.3
yarg==0.1.9
yarl==1.6.3
